Updated installation script for jetpack 4.5

+++ install.sh
@@ -24,7 +24,7 @@ sudo apt-get update
 sudo apt install -y python3-pip python3-pil python3-smbus python3-matplotlib cmake
 sudo -H pip3 install --upgrade pip
 sudo -H pip3 install flask
-sudo -H pip3 install --upgrade numpy
+#sudo -H pip3 install --upgrade numpy # we need numpy 1.16.1 for tensorflow 
 
 # Install jtop
 echo "\e[100m Install jtop \e[0m"
@@ -34,24 +34,31 @@ sudo -H pip install jetson-stats
 # Install the pre-built TensorFlow pip wheel
 echo "\e[48;5;202m Install the pre-built TensorFlow pip wheel \e[0m"
 sudo apt-get update
-sudo apt-get install -y libhdf5-serial-dev hdf5-tools libhdf5-dev zlib1g-dev zip libjpeg8-dev
+sudo apt-get install -y libhdf5-serial-dev hdf5-tools libhdf5-dev zlib1g-dev zip libjpeg8-dev liblapack-dev libblas-dev gfortran
 sudo apt-get install -y python3-pip
-sudo -H pip3 install -U pip setuptools
-sudo -H pip3 install -U numpy grpcio absl-py py-cpuinfo psutil portpicker six mock requests gast h5py astor termcolor protobuf keras-applications keras-preprocessing wrapt google-pasta
-sudo -H pip3 install --pre --extra-index-url https://developer.download.nvidia.com/compute/redist/jp/v43 tensorflow==1.15.2+nv20.3
+sudo -H pip3 install -U pip testresources setuptools==49.6.0 
+sudo -H pip3 install -U numpy==1.16.1 future==0.18.2 mock==3.0.5 h5py==2.10.0 keras_preprocessing==1.1.1 keras_applications==1.0.8 gast==0.2.2 futures protobuf pybind11
+sudo -H pip3 install --pre --extra-index-url https://developer.download.nvidia.com/compute/redist/jp/v45 'tensorflow<2'
 
 # Install the pre-built PyTorch pip wheel 
 echo "\e[45m Install the pre-built PyTorch pip wheel  \e[0m"
 cd
-wget -N https://nvidia.box.com/shared/static/9eptse6jyly1ggt9axbja2yrmj6pbarc.whl -O torch-1.6.0-cp36-cp36m-linux_aarch64.whl 
+if [ ! -f torch-1.6.0-cp36-cp36m-linux_aarch64.whl  ]; then
+    wget -N https://nvidia.box.com/shared/static/9eptse6jyly1ggt9axbja2yrmj6pbarc.whl -O torch-1.6.0-cp36-cp36m-linux_aarch64.whl 
+fi
+
 sudo apt-get install -y python3-pip libopenblas-base libopenmpi-dev 
 sudo -H pip3 install Cython
-sudo -H pip3 install numpy torch-1.6.0-cp36-cp36m-linux_aarch64.whl
+#sudo -H pip3 install numpy torch-1.6.0-cp36-cp36m-linux_aarch64.whl # we need numpy 1.16.1 for tensorflow 
+sudo -H pip3 install torch-1.6.0-cp36-cp36m-linux_aarch64.whl
 
 # Install torchvision package
 echo "\e[45m Install torchvision package \e[0m"
+sudo apt-get install -y libavformat-dev libswscale-dev
 cd
-git clone https://github.com/pytorch/vision torchvision
+if [ ! -d torchvision  ]; then
+    git clone https://github.com/pytorch/vision torchvision 
+fi
 cd torchvision
 git checkout tags/v0.7.0
 sudo -H python3 setup.py install
@@ -71,11 +78,20 @@ echo "\e[48;5;172m Install traitlets \e[0m"
 #sudo -H python3 -m pip install git+https://github.com/ipython/traitlets@master
 sudo python3 -m pip install git+https://github.com/ipython/traitlets@dead2b8cdde5913572254cf6dc70b5a6065b86f8
 
+# Install newer version of Node.js for the latest Jupyter Lab
+echo "\e[48;5;172m Install Node.js \e[0m"
+sudo apt install -y curl
+curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.3/install.sh | bash
+export NVM_DIR="$HOME/.nvm" # nvm would be picked when terminal is restarted, but we are still installing packages
+[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
+[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+nvm install 12.18.1
+curl -sL https://deb.nodesource.com/setup_12.x | sudo -E bash -
+sudo apt-get install -y nodejs
+
 # Install Jupyter Lab
 echo "\e[48;5;172m Install Jupyter Lab \e[0m"
-sudo apt install -y curl
-curl -sL https://deb.nodesource.com/setup_10.x | sudo -E bash -
-sudo apt install -y nodejs libffi-dev
+sudo apt install -y libffi-dev
 sudo -H pip3 install jupyter jupyterlab
 sudo -H jupyter labextension install @jupyter-widgets/jupyterlab-manager
 
@@ -83,7 +99,7 @@ jupyter lab --generate-config
 python3 -c "from notebook.auth.security import set_password; set_password('$password', '$HOME/.jupyter/jupyter_notebook_config.json')"
 
 # fix for Traitlet permission error
-sudo chown -R jetson:jetson ~/.local/share/
+sudo chown -R $USER:$USER ~/.local/share/
 
 # Install jetcard
 echo "\e[44m Install jetcard \e[0m"
@@ -146,7 +162,9 @@ cd
 sudo apt-get install libssl1.0-dev
 #sudo apt-get install nodejs-dev node-gyp libssl1.0-dev
 #sudo apt-get install npm
-git clone https://github.com/jaybdub/jupyter_clickable_image_widget
+if [ ! -d jupyter_clickable_image_widget  ]; then
+    git clone https://github.com/jaybdub/jupyter_clickable_image_widget
+fi
 cd jupyter_clickable_image_widget
 git checkout no_typescript
 sudo -H pip3 install -e .
